Complete two python classes. It's working well. I'm trying to modify the impinj configuration

+++ impinj_t1.h
@@ -91,12 +91,12 @@ private:
 
     int getOnePhaseAngle(
       LLRP::CImpinjRFPhaseAngle  *pRfPhase,
-      double               *out);
+      unsigned int               *out);
 
     int
     getOnePeakRSSI (
       LLRP::CImpinjPeakRSSI      *pPeakRSSI,
-      double               *out);
+      unsigned int               *out);
 
     int
     getOneTimestamp (
